+++ src/extension.ts
@@ -45,11 +45,11 @@ function formatDocument(fmtCmds: string[], doc: vscode.TextDocument): string {
 		GOPATH: goPath,
 		PATH: goPath + "/bin:"+ process.env.PATH,
 	};
-
+	const text = doc.getText();
 	const outBuf = cp.spawnSync(fmtCmds[0],fmtCmds.slice(1), {
 		 env,
 		 encoding: 'utf8',
-		 input: doc.getText(),
+		 input: text,
 	});
 
 	if (outBuf.error || outBuf.status !== 0 || outBuf.stderr.length > 0) {
@@ -58,17 +58,27 @@ function formatDocument(fmtCmds: string[], doc: vscode.TextDocument): string {
 		return "";
 	}
 
-	return outBuf.stdout;
+	return outBuf.stdout !== text ? outBuf.stdout : "";
 }
 
-function grabyrightOnSave(fmtCmds: string[], e: vscode.TextDocumentWillSaveEvent) {
+function formatDocumentOnSave(e: vscode.TextDocumentWillSaveEvent) {
 	const doc = e.document;
 	if (doc.languageId !== "go") {
 		return;
 	}
 
-	e.waitUntil(new Promise((resolve, reject) => {
-		const formatted = formatDocument(fmtCmds, doc);
+	const config = vscode.workspace.getConfiguration("goCustomFormat");
+	if (!config["onSave"]) {
+		return;
+	}
+
+	const activeDoc = vscode.window.activeTextEditor ? vscode.window.activeTextEditor.document : undefined;
+	if (activeDoc !== doc) {
+		return;
+	}
+
+	e.waitUntil(new Promise((resolve) => {
+		const formatted = formatDocument(config["fmtCmds"], doc);
 		if (formatted.length === 0 || formatted === doc.getText()) {
 			resolve([]);
 			return;
@@ -81,40 +91,19 @@ function grabyrightOnSave(fmtCmds: string[], e: vscode.TextDocumentWillSaveEvent
 	}));
 }
 
-function prepareOnSave(curSub: vscode.Disposable | undefined, subscriptions: vscode.Disposable[]): vscode.Disposable | undefined {
-	const config = vscode.workspace.getConfiguration('goCustomFormat', vscode.window.activeTextEditor ? vscode.window.activeTextEditor.document.uri : null);
-	if (!curSub && config['onSave']) {
-		return vscode.workspace.onWillSaveTextDocument(
-			e => grabyrightOnSave(config["fmtCmds"], e),
-			null,
-			subscriptions,
-		);
-	}
-
-	if (curSub && !config['onSave']) {
-		curSub.dispose();
-		return undefined;
-	}
-
-	return curSub;
-}
-
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(ctx: vscode.ExtensionContext) {
 		// Use the console to output diagnostic information (console.log) and errors (console.error)
 		// This line of code will only be executed once when your extension is activated
-		console.log('Congratulations, your extension "goCustomFormat" is now active!');
 		ctx.subscriptions.push(vscode.commands.registerTextEditorCommand(
 			"extension.goCustomFormat",
 			runCustomFormatCmd,
 		));
 
-		let onSaveSubscription = prepareOnSave(undefined, ctx.subscriptions);
-		ctx.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
-			// config updated
-			onSaveSubscription = prepareOnSave(onSaveSubscription, ctx.subscriptions);
-		}));
+		ctx.subscriptions.push(vscode.workspace.onWillSaveTextDocument(
+			formatDocumentOnSave,
+		));
 }
 
 // this method is called when your extension is deactivated