+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.0] - 2018-07-21
+### Changed
+- Only format the active document on TextDocumentWillSave events.
+
 ## [0.0.1] - 2018-07-21
 ### Added
-- Command `Run custom format`
-- Configs `goCustomFormat.onSave` and `goCustomFormat.fmtCmds`
+- Command `Run custom format`.
+- Configs `goCustomFormat.onSave` and `goCustomFormat.fmtCmds`.